Add sum2 phase
