+++ rust/src/coordinator_async/message_processing.rs
@@ -1,7 +1,7 @@
 use crate::{
     coordinator::RoundSeed,
     crypto::ByteObject,
-    message::{SumOwned, UpdateOwned},
+    message::{Sum2Owned, SumOwned, UpdateOwned},
     ParticipantPublicKey,
     ParticipantTaskSignature,
     PetError,
@@ -28,22 +28,25 @@ pub struct UpdateValidationData {
     pub seed: RoundSeed,
 }
 
-// A sink to collect the results of the MessageValidator tasks.
+// A sink to collect the results of the MessageHandler tasks.
 pub struct MessageSink {
     // The minimum number of successfully validated messages.
     min_messages: usize,
     // A counter that is incremented when a message has been successfully validated.
     successful_messages: usize,
-    // The channel receiver that receives the results of the MessageValidator tasks.
+    // The channel receiver that receives the results of the MessageHandler tasks.
     sink_rx: UnboundedReceiver<Result<(), PetError>>,
     // The minimum duration to wait.
     min_duration: Duration,
+    // Message collection time out.
+    max_duration: Duration,
 }
 
 impl MessageSink {
     pub fn new(
         min_messages: usize,
         min_duration: Duration,
+        max_duration: Duration,
     ) -> (UnboundedSender<Result<(), PetError>>, Self) {
         let (success_tx, sink_rx) = unbounded_channel();
         (
@@ -53,6 +56,7 @@ impl MessageSink {
                 successful_messages: 0,
                 sink_rx,
                 min_duration,
+                max_duration,
             },
         )
     }
@@ -63,21 +67,20 @@ impl MessageSink {
             mut successful_messages,
             mut sink_rx,
             min_duration,
+            max_duration,
         } = self;
-        println!("Sink collect");
-
-        // Collect the results of the MessageValidator tasks. The collect future will be
+        // Collect the results of the MessageHandler tasks. The collect future will be
         // successfully resolved when the minimum duration has been waited and when the minimum
         // number of successful validated messages has been reached.
-        // The first failed MessageValidator result causes the collection to be canceled. In this
+        // The first failed MessageHandler result causes the collection to be canceled. In this
         // case the collect future will be resolved with an error.
-        let wait_min_duration = async move {
+        let min_collection_duration = async move {
             tokio::time::delay_for(min_duration).await;
-            println!("Min time frame complete");
+            println!("waited min collection duration");
             Ok::<(), PetError>(())
         };
 
-        let collection = async move {
+        let message_collection = async move {
             loop {
                 let _message = sink_rx
                     .recv()
@@ -94,21 +97,33 @@ impl MessageSink {
             }
         };
 
-        tokio::try_join!(wait_min_duration, collection).map(|_| ())
+        let mut max_collection_duration = tokio::time::delay_for(max_duration);
+
+        tokio::select! {
+            collection_result = async {
+                tokio::try_join!(min_collection_duration, message_collection).map(|_| ())
+            }=> {
+                collection_result
+            }
+            _ = &mut max_collection_duration => {
+                println!("message collection timed out!");
+                Err::<(), PetError>(PetError::InvalidMessage)
+            }
+        }
     }
 }
 
-pub struct MessageValidator {
+pub struct MessageHandler {
     // A sender through which the result of the massage validation is sent.
     sink_tx: UnboundedSender<Result<(), PetError>>,
     // We will never send anything on this channel.
-    // We use this channel to keep track of which MessageValidator tasks are still alive.
+    // We use this channel to keep track of which MessageHandler tasks are still alive.
     _cancel_complete_tx: Sender<()>,
     // The channel receiver that receives the cancel notification.
     notify_cancel: broadcast::Receiver<()>,
 }
 
-impl MessageValidator {
+impl MessageHandler {
     pub fn new(
         sink_tx: UnboundedSender<Result<(), PetError>>,
         cancel_complete_tx: Sender<()>,
@@ -125,11 +140,11 @@ impl MessageValidator {
         self,
         message_validation_fut: impl Future<Output = Result<(), PetError>>,
     ) {
-        // Extract all fields of the MessageValidator struct. This is necessary to bypass borrow
+        // Extract all fields of the MessageHandler struct. This is necessary to bypass borrow
         // issues in the tokio::select macro.
         // It is important to extract _cancel_complete_tx as well, otherwise the channel
         // will be dropped too early.
-        let MessageValidator {
+        let MessageHandler {
             sink_tx,
             _cancel_complete_tx,
             mut notify_cancel,
@@ -145,7 +160,7 @@ impl MessageValidator {
 }
 
 // Sum message validator
-impl MessageValidator {
+impl MessageHandler {
     /// Validate and handle a sum message.
     pub async fn handle_sum_message(
         self,
@@ -180,7 +195,7 @@ impl MessageValidator {
 }
 
 // Update message validator
-impl MessageValidator {
+impl MessageHandler {
     /// Validate and handle an update message.
     pub async fn handle_update_message(
         self,
@@ -234,5 +249,26 @@ impl MessageValidator {
         } else {
             Err(PetError::InvalidMessage)
         }
+    }
+}
+
+// Sum2 message validator
+impl MessageHandler {
+    /// Validate and handle an update message.
+    pub async fn handle_sum2_message(
+        self,
+        coordinator_state: Arc<SumValidationData>,
+        pk: ParticipantPublicKey,
+        message: Sum2Owned,
+    ) {
+        let message_validation_fut = async {
+            // if !self.sum_dict.contains_key(&pk) {
+            //     return Err(PetError::InvalidMessage);
+            // }
+            Self::validate_sum_task(&coordinator_state, &pk, &message.sum_signature)
+            // async call to Redis
+            //self.add_mask(&pk, message.mask).unwrap();
+        };
+        self.handle_message(message_validation_fut).await;
     }
 }
