+++ rust/src/lib.rs
@@ -13,7 +13,6 @@ pub mod coordinator_async;
 pub mod crypto;
 pub mod mask;
 pub mod message;
-pub mod message_processing;
 pub mod participant;
 pub mod service;
